Introduce key deletion for char seq.
This commit will introduce a feature which will enable users to
delete keys which match a given char sequence.
To achieve that the user only has to specify the char sequence
followed by an asterics in the key list parameter.
E.g. *ExifT will delete keys like: ExifTool, ExifTime etc.

This feature might need some more work and optimisation as its
using a n*n notation which might be really unefficient depending
on the list of char sequences the user specifies.

It might also be wise to get back to the delete_keys_with_str_seq
function for refactoring and keeping seperation of concerns.