Add feature to specify keys for mediainfo info
This commit will introduce the same behaviour we use for
the exif tool. The user has to add a file with keys to call the
specific keys from the mediainfo output. At the moment the exif tool
and the mediainfo tool use the same parameter, this can easily be
changed if necessary.