Fix issue where mediainfo wasnt displayed properly
This commit should fix an issue which was related to the one
solved by the commit before for the exif_log. This commit uses
the same technique to overcome the problem in mediainfo.
In addition to that the type is now a parent key for the keys
to determine where the fieldkeys and the values belong to.