Fix field key stats
This commit will fix two problems. Firstly the one that the
fieldkey statistic is printed on every iteration, and secondly
an issue which occured causing the wrong args variable.
In a future commit, it might be worth checking for better
if clauses, as this adds two if clauses for the same args variable.