Finalise the Siegfried merge and start w exif
This commit finalised the merge functionality for Siegfried.
This should work fine by now, additional tests need to be written.
Also started with merging the exif log however this comes with a
few problems.
1. I am not sure if its smart that I have to call the siegfried
file info again - and if so - then we will desperately need to
change how the main function works. E.g.: If there is an exif log
just run the merge_exif function and dont run the merge sf function
cause this would be repetetive, cause the merging happens anyway.

2. We need to implement the replace non value function, and even
worse the functionality which selects the field keys specified in the
key list. We probably want to add those to the add_to_db function to have the
cleaned up data in the db and not the raw data. Could argue both ways.