Introduce functional version of mediainfo merge
This desperately needs more pytests as well as tests to
check if the memory usage is optimised. Probably lots of potential
to optimise things here. Selecting Path as primary key instead of
pulling the whole db.
Another problem might be that merged_log_files in main.py might
stores the merged_log file over the whole function while this would
be superficial as the information are pulled in the specific functions
again anyways.