Introduce 2 fixes
First fix is regarding the occurence of keys where the if clause
wasn't working as expected. Specifying the if clause to be True
fixes it now.

Second fix is regarding the counter which was not dependent on the
update to db but was falsely dependend on the iteration of files in
the log.
After this commit the counter is displaying how many entries are
uploaded to the db as stated and not longer how many files are in
a log. If needed we could add that information too.