Introduce If clause to ensure unique paths
This commit fixes in issue which was caused by duplicate files.
As we use the file path as unique PK, we have to ensure that every
file is only uploaded once to the DB. A simple if clause which checks
if the file path is already in the DB will fix that issue.
It has to be investigated whether this has a negative effect on
the performance.